+++ includes/admin/tools/config/functions-config.php
@@ -94,7 +94,7 @@ function update_config_post(int $postID, array $updateValues = []): string {
         foreach ($new as $key => $value) {
 
             //TODO
-            if(in_array($key, ['admin_columns', 'metadata', 'archive', 'archive_filter', 'authors', 'creators', 'literature', 'terms', 'external', 'lod'])){
+            if(in_array($key, ['admin_columns', 'metadata', 'archive', 'archive_filter', 'authors', 'creators', 'literature', 'terms', 'external', 'lod', 'relations', 'related_content', 'translators', 'editors'])){
                 $current[$key] = normalize_option_value($value, $current[$key] ?? []);
             }
             elseif (is_array($value)) {